Melhorias e novas implementações: - Alterado a extensão de rotas Flask-RESTful para a flask-smorest - Adicionado a extensão marshmallow para serialização de objeto - Melhorias na lógica do serviço de produto

+++ flask-microservices-products/app/services/product_service.py
@@ -1,3 +1,4 @@
+from app.errors.resource_not_found_error import ResourceNotFoundError
 from app.repositories.product_repository import ProductRepository
 
 
@@ -9,10 +10,31 @@ class ProductService:
         return self.product_repository.get_products()
 
     def get_product_by_id(self, id):
+        product = self.product_repository.get_product_by_id(id)
+
+        if not product:
+            raise ResourceNotFoundError(f'Product not found with identifier {id}')
+
         return self.product_repository.get_product_by_id(id)
 
-    def save_product(self, product):
+    def product_exists_by_id(self, id):
+        return self.product_repository.product_exists_by_id(id)
+
+    def create_product(self, product):
+        return self.product_repository.save_product(product)
+
+    def update_product(self, id, product):
+        if not self.product_exists_by_id(id):
+            raise ResourceNotFoundError(f'Product not found with identifier {id}')
+
+        product.id = id
+
         return self.product_repository.save_product(product)
 
     def delete_product_by_id(self, id):
+        product = self.product_repository.get_product_by_id(id)
+
+        if not product:
+            raise ResourceNotFoundError(f'Product not found with identifier {id}')
+
         self.product_repository.delete_product_by_id(id)
